+++ resources/views/livewire/permissions/show-group-members.blade.php
@@ -10,7 +10,7 @@
             <h5>Members of {{ $group->name }}</h5>
             <label for="groupMemberSearch" class="float-end mx-2">
                 <input id="groupMemberSearch" type="search" wire:model="search" class="form-control"
-                       placeholder="Search..." />
+                       placeholder="Search..."/>
             </label>
         </div>
         <div class="card-body border-0 shadow table-responsive">
@@ -20,7 +20,9 @@
                     <th>Username</th>
                     <th>Server</th>
                     <th>Expired</th>
-                    <th>Actions</th>
+                    @can('edit_permissions')
+                        <th>Actions</th>
+                    @endcan
                 </tr>
                 </thead>
                 <tbody>
@@ -39,18 +41,21 @@
                         <td>{{ $member->permissionPlayer->name }}</td>
                         <td>{{ $server }}</td>
                         <td>{{ $expires }}</td>
-                        <td>
-                            {{--<button type="button" style="background: transparent; border: none;" data-mdb-toggle="modal"
-                                    data-mdb-target="#editGroupModal"
-                                    wire:click="editGroup({{$group->id}})">
-                                <i class="material-icons text-warning">edit</i>
-                            </button>--}}
-                            <button type="button" style="background: transparent; border: none;" data-mdb-toggle="modal"
-                                    data-mdb-target="#deleteGroupMemberModal"
-                                    wire:click="deleteGroupMember({{ $member->id }}, '{{ $member->permissionPlayer->uuid }}')">
-                                <i class="material-icons text-danger">delete</i>
-                            </button>
-                        </td>
+                        @can('edit_permissions')
+                            <td>
+                                {{--<button type="button" style="background: transparent; border: none;" data-mdb-toggle="modal"
+                                        data-mdb-target="#editGroupModal"
+                                        wire:click="editGroup({{$group->id}})">
+                                    <i class="material-icons text-warning">edit</i>
+                                </button>--}}
+                                <button type="button" style="background: transparent; border: none;"
+                                        data-mdb-toggle="modal"
+                                        data-mdb-target="#deleteGroupMemberModal"
+                                        wire:click="deleteGroupMember({{ $member->id }}, '{{ $member->permissionPlayer->uuid }}')">
+                                    <i class="material-icons text-danger">delete</i>
+                                </button>
+                            </td>
+                        @endcan
                     </tr>
                 @endforeach
                 </tbody>
