+++ resources/views/livewire/permissions/show-group-parents.blade.php
@@ -17,25 +17,30 @@
                 <thead>
                 <tr>
                     <th>Parent Group</th>
-                    <th>Actions</th>
+                    @can('edit_permissions')
+                        <th>Actions</th>
+                    @endcan
                 </tr>
                 </thead>
                 <tbody>
                 @foreach($parents as $parent)
                     <tr>
                         <td>{{ $parent->parentGroup->name }}</td>
-                        <td>
-                            {{--<button type="button" style="background: transparent; border: none;" data-mdb-toggle="modal"
-                                    data-mdb-target="#editGroupModal"
-                                    wire:click="editGroup({{$group->id}})">
-                                <i class="material-icons text-warning">edit</i>
-                            </button>--}}
-                            <button type="button" style="background: transparent; border: none;" data-mdb-toggle="modal"
-                                    data-mdb-target="#deleteGroupParentModal"
-                                    wire:click="deleteGroupParent({{ $parent->id }}, {{ $parent->parentgroupid }})">
-                                <i class="material-icons text-danger">delete</i>
-                            </button>
-                        </td>
+                        @can('edit_permissions')
+                            <td>
+                                {{--<button type="button" style="background: transparent; border: none;" data-mdb-toggle="modal"
+                                        data-mdb-target="#editGroupModal"
+                                        wire:click="editGroup({{$group->id}})">
+                                    <i class="material-icons text-warning">edit</i>
+                                </button>--}}
+                                <button type="button" style="background: transparent; border: none;"
+                                        data-mdb-toggle="modal"
+                                        data-mdb-target="#deleteGroupParentModal"
+                                        wire:click="deleteGroupParent({{ $parent->id }}, {{ $parent->parentgroupid }})">
+                                    <i class="material-icons text-danger">delete</i>
+                                </button>
+                            </td>
+                        @endcan
                     </tr>
                 @endforeach
                 </tbody>
@@ -45,10 +50,12 @@
             </div>
         </div>
     </div>
-    <div class="p-4">
-        <button type="button" class="btn btn-primary" data-mdb-toggle="modal" data-mdb-target="#addGroupParentModal"
-                wire:click="addGroupParent">
-            <i style="font-size: 18px !important;" class="material-icons">add</i> Add Group
-        </button>
-    </div>
+    @can('edit_permissions')
+        <div class="p-4">
+            <button type="button" class="btn btn-primary" data-mdb-toggle="modal" data-mdb-target="#addGroupParentModal"
+                    wire:click="addGroupParent">
+                <i style="font-size: 18px !important;" class="material-icons">add</i> Add Group
+            </button>
+        </div>
+    @endcan
 </div>
