+++ resources/views/livewire/permissions/show-groups.blade.php
@@ -10,7 +10,7 @@
             <h5>Permission Groups</h5>
             <label for="groupSearch" class="float-end mx-2">
                 <input id="groupSearch" type="search" wire:model="search" class="form-control"
-                       placeholder="Search..." />
+                       placeholder="Search..."/>
             </label>
         </div>
         <div class="card-body border-0 shadow table-responsive">
@@ -37,16 +37,20 @@
                                     wire:click="showGroup({{$group->id}})">
                                 <i class="material-icons text-info">info</i>
                             </button>
-                            <button type="button" style="background: transparent; border: none;" data-mdb-toggle="modal"
-                                    data-mdb-target="#editGroupModal"
-                                    wire:click="editGroup({{$group->id}})">
-                                <i class="material-icons text-warning">edit</i>
-                            </button>
-                            <button type="button" style="background: transparent; border: none;" data-mdb-toggle="modal"
-                                    data-mdb-target="#deleteGroupModal"
-                                    wire:click="deleteGroup({{ $group->id }})">
-                                <i class="material-icons text-danger">delete</i>
-                            </button>
+                            @can('edit_permissions')
+                                <button type="button" style="background: transparent; border: none;"
+                                        data-mdb-toggle="modal"
+                                        data-mdb-target="#editGroupModal"
+                                        wire:click="editGroup({{$group->id}})">
+                                    <i class="material-icons text-warning">edit</i>
+                                </button>
+                                <button type="button" style="background: transparent; border: none;"
+                                        data-mdb-toggle="modal"
+                                        data-mdb-target="#deleteGroupModal"
+                                        wire:click="deleteGroup({{ $group->id }})">
+                                    <i class="material-icons text-danger">delete</i>
+                                </button>
+                            @endcan
                         </td>
                     </tr>
                 @endforeach
@@ -57,10 +61,12 @@
             </div>
         </div>
     </div>
-    <div class="p-4">
-        <button type="button" class="btn btn-primary" data-mdb-toggle="modal" data-mdb-target="#addGroupModal"
-                wire:click="addGroup">
-            <i style="font-size: 18px !important;" class="material-icons">add</i> Add Group
-        </button>
-    </div>
+    @can('edit_permissions')
+        <div class="p-4">
+            <button type="button" class="btn btn-primary" data-mdb-toggle="modal" data-mdb-target="#addGroupModal"
+                    wire:click="addGroup">
+                <i style="font-size: 18px !important;" class="material-icons">add</i> Add Group
+            </button>
+        </div>
+    @endcan
 </div>
